apps/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0
supabase>=2.0.0
openai>=1.0.0
PyGithub>=2.0.0
PyJWT>=2.8.0
notion-client>=2.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
langchain-text-splitters>=0.0.1
tiktoken>=0.5.0

# Dev dependencies
pytest>=8.0.0
